add logic for consumption history, add better front-end look

+++ Sourcecode/src/main/java/com/koi/FA24_SE1854_SWP391_G6_KoiCareHome/model/ConsumeFoodHistory.java
@@ -27,7 +27,7 @@ public class ConsumeFoodHistory {
     private int foodID;
 
     @Column(name = "Quantity", nullable = false)
-    private float quantity;
+    private int quantity;
 
     @Column(name = "Description", nullable = true)
     private String description;
